First commit with multithreading. To test on working env.
